requirements.txt
fastapi
uvicorn
pymongo
google-cloud-aiplatform
python-dotenv
